Refactor location import & export plugins
