+++ src/PluginManager/include/PluginManager/Location/LocationImportIntf.h
@@ -27,7 +27,7 @@
 
 #include <QtPlugin>
 
-#include "PluginIntf.h"
+#include "../PluginIntf.h"
 
 class LocationImportIntf : public PluginIntf
 {
