+++ src/PluginManager/src/Location/BasicLocationImportDialog.ui
